pauz-backend/requirements.txt
fastapi
uvicorn
google-auth-oauthlib
google-api-python-client
pydantic
python-dotenv
reportlab
boto3
lm-raindrop-integrations
elevenlabs
python-jose[cryptography]
passlib[bcrypt]
sqlmodel
